Created separate modules for creation of each entity
